权限管理

+++ src/main/java/com/djyz/domain/ClothesOrderState.java
@@ -4,7 +4,9 @@ import lombok.Data;
 
 @Data
 public class ClothesOrderState {
-    private Integer cosId;
+    private Long cosId;
+
     private String cosName;
+
 
 }
